Move filter WithRoutine to top
As WithRoutine changes the serving behavior with early return for
Watch requests, which affect other fitlers to work properly. See
https://github.com/kubernetes/kubernetes/pull/125626 &
https://github.com/kubernetes/kubernetes/issues/126466. So this commit
moves it to top of filter chain so no other filters would be impacted
by early return.
