+++ staging/src/k8s.io/apiserver/pkg/server/httplog/httplog.go
@@ -31,7 +31,6 @@ import (
 	"k8s.io/apiserver/pkg/endpoints/metrics"
 	"k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/endpoints/responsewriter"
-	"k8s.io/apiserver/pkg/server/routine"
 	"k8s.io/klog/v2"
 )
 
@@ -126,26 +125,10 @@ func withLogging(handler http.Handler, stackTracePred StacktracePred, shouldLogR
 		rl := newLoggedWithStartTime(req, w, startTime)
 		rl.StacktraceWhen(stackTracePred)
 		req = req.WithContext(context.WithValue(ctx, respLoggerContextKey, rl))
-
-		var logFunc func()
-		logFunc = rl.Log
-		defer func() {
-			if logFunc != nil {
-				logFunc()
-			}
-		}()
+		defer rl.Log()
 
 		w = responsewriter.WrapForHTTP1Or2(rl)
 		handler.ServeHTTP(w, req)
-
-		// We need to ensure that the request is logged after it is processed.
-		// In case the request is executed in a separate goroutine created via
-		// WithRoutine handler in the handler chain (i.e. above handler.ServeHTTP()
-		// would return request is completely responsed), we want the logging to
-		// happen in that goroutine too, so we append it to the task.
-		if routine.AppendTask(ctx, &routine.Task{Func: rl.Log}) {
-			logFunc = nil
-		}
 	})
 }
